修改测试bug http://172.17.96.87:8080/v3/api/info/measure_region_rel_v2/?timestamp=0&project_id=6&last_id=0

+++ longjian-measure-app/src/main/java/com/longfor/longjian/measure/app/vo/proPaintAreaManageVo/RelVo.java
@@ -29,9 +29,9 @@ public class RelVo {
     /**
      * 更新时间
      */
-    private Long update_at;
+    private int update_at;
     /**
      * 删除时间
      */
-    private Long delete_at;
+    private int delete_at;
 }
